test: add reader adapter test
