+++ internal/utils/utils.go
@@ -3,7 +3,6 @@ package utils
 import (
 	"bufio"
 	"debug/pe"
-	"errors"
 	"fmt"
 	"io"
 	"iter"
@@ -119,24 +118,6 @@ func PrefixToPath(s string) (string, error) {
 		i += 3
 	}
 	return string(p), nil
-}
-
-type ReaderAtAdapter struct {
-	readerAt io.ReaderAt
-	offset   int64
-}
-
-func NewReaderAtAdapter(readerAt io.ReaderAt) *ReaderAtAdapter {
-	return &ReaderAtAdapter{readerAt: readerAt}
-}
-
-func (r *ReaderAtAdapter) Read(p []byte) (n int, err error) {
-	n, err = r.readerAt.ReadAt(p, r.offset)
-	r.offset += int64(n)
-	if errors.Is(err, io.EOF) && n > 0 {
-		return n, nil
-	}
-	return
 }
 
 func IsJson(s io.ReaderAt) bool {
